requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
async_generator=1.10=py_0
attrs=21.4.0=pyhd8ed1ab_0
brotlipy=0.7.0=py310h1961e1f_1004
bzip2=1.0.8=h0d85af4_4
ca-certificates=2022.5.18=h033912b_0
certifi=2022.5.18=py310h2ec42d9_0
cffi=1.15.0=py310hcc37b68_0
cryptography=37.0.2=py310h52c3658_0
h11=0.13.0=pyhd8ed1ab_1
icu=70.1=h96cf925_0
idna=3.3=pyhd8ed1ab_0
libblas=3.9.0=14_osx64_openblas
libcblas=3.9.0=14_osx64_openblas
libcxx=14.0.3=hc203e6f_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=14_osx64_openblas
libopenblas=0.3.20=openmp_hb3cd9ec_0
libxml2=2.9.14=h08a9926_0
libxslt=1.1.33=h5bff336_4
libzlib=1.2.11=h6c3fc93_1014
llvm-openmp=14.0.3=ha654fa7_0
lxml=4.8.0=py310h1961e1f_3
ncurses=6.3=h96cf925_1
numpy=1.22.3=py310he1f95b8_2
openssl=1.1.1o=hfe4f2af_0
outcome=1.1.0=pyhd8ed1ab_0
pandas=1.4.2=py310h514ec25_1
pycparser=2.21=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_0
pysocks=1.7.1=py310h2ec42d9_5
python=3.10.4=h8b4d769_0_cpython
python-chromedriver-binary=102.0.5005.27.0=py310h2ec42d9_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pytz=2022.1=pyhd8ed1ab_0
readline=8.1=h05e3726_0
selenium=4.1.0=pyhd8ed1ab_0
setuptools=62.3.1=py310h2ec42d9_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.2.0=py310h2ec42d9_3
sortedcontainers=2.4.0=pyhd8ed1ab_0
sqlite=3.38.5=hd9f0692_0
tk=8.6.12=h5dbffcc_0
trio=0.20.0=py310h2ec42d9_1
trio-websocket=0.9.2=pyhd8ed1ab_0
typing_extensions=4.2.0=pyha770c72_1
tzdata=2022a=h191b570_0
urllib3=1.26.9=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
wsproto=1.1.0=py310h2ec42d9_1
xz=5.2.5=haf1e3a3_1
zlib=1.2.11=h6c3fc93_1014
